Add files via upload
Tic-tac-toe is a simple, two-player game that, if played optimally by both players, will always result in a tie. The game is also called noughts and crosses or Xs and Os.
I did it in C++ language , later on I will implement it by using some libraries so that a single player can play with Computer.